FIX(Get most of recently ignored tests working again under Windows)
