+++ src/main/java/mmarquee/automation/controls/AutomationStatusBar.java
@@ -43,6 +43,17 @@ public class AutomationStatusBar extends AutomationContainer {
      * Constructor for AutomationStatusBar
      * @param element The underlying element
      * @param pattern The range pattern
+     * @throws AutomationException Automation library error
+     * @throws PatternNotFoundException Did not find the pattern
+     */
+    AutomationStatusBar(AutomationElement element, ItemContainer pattern) throws PatternNotFoundException, AutomationException{
+        super(element, pattern);
+    }
+
+    /**
+     * Constructor for AutomationStatusBar
+     * @param element The underlying element
+     * @param pattern The range pattern
      * @param instance The automation instance
      * @throws AutomationException Automation library error
      * @throws PatternNotFoundException Did not find the pattern
