+++ src/test/java/mmarquee/automation/controls/ribbon/AutomationNUIPaneTest.java
@@ -15,33 +15,21 @@
  */
 package mmarquee.automation.controls.ribbon;
 
-import com.sun.jna.Memory;
-import com.sun.jna.Native;
-import com.sun.jna.Pointer;
-import com.sun.jna.ptr.PointerByReference;
 import mmarquee.automation.AutomationElement;
 
 import static org.junit.Assert.*;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.when;
 
-import mmarquee.automation.ElementNotFoundException;
 import mmarquee.automation.UIAutomation;
 import mmarquee.automation.pattern.ItemContainer;
 import mmarquee.automation.uiautomation.IUIAutomation;
-import mmarquee.automation.uiautomation.IUIAutomationElement3;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 
-import java.util.ArrayList;
-import java.util.List;
-
 /**
- * Created by Mark Humphreys on 01/12/2016.
+ * Created by Mark Humphreys on 28/11/2016.
  *
  * Tests for the NUIPane control.
  */
@@ -73,68 +61,5 @@ public class AutomationNUIPaneTest {
         String name = pane.getName();
 
         assertTrue(name.isEmpty());
-    }
-
-    @Test(expected = ElementNotFoundException.class)
-    @Ignore("Requires windows")
-    public void testGetNetUIHWND_Throws_Exception_When_NetUIHWND_Not_Found() throws Exception {
-        AutomationElement element = Mockito.mock(AutomationElement.class);
-
-        when(element.getClassName()).thenReturn(AutomationNUIPane.CLASS_NAME);
-        List<AutomationElement> collection = new ArrayList<>();
-
-        when(element.findAll(any(), any())).thenReturn(collection);
-
-        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
-        UIAutomation instance = new UIAutomation(mocked_automation);
-
-        ItemContainer container = Mockito.mock(ItemContainer.class);
-
-        AutomationNUIPane pane = new AutomationNUIPane(element, container, instance);
-
-        pane.getNetUIHWND(0);
-
-        Mockito.verify(element, atLeastOnce()).findAll(any(), any());
-    }
-
-    @Test
-    @Ignore("Requires windows")
-    public void testGetNetUIHWND_When_NetUIHWND_Is_Found() throws Exception {
-        AutomationElement element = Mockito.mock(AutomationElement.class);
-
-        when(element.getClassName()).thenReturn(AutomationNUIPane.CLASS_NAME);
-        List<AutomationElement> collection = new ArrayList<>();
-
-        IUIAutomationElement3 elem = Mockito.mock(IUIAutomationElement3.class);
-
-        Mockito.when(elem.getCurrentClassName(any())).thenAnswer(
-            invocation -> {
-                Object[] args = invocation.getArguments();
-                PointerByReference reference = (PointerByReference) args[0];
-
-                String value = "NetUIHWND";
-                Pointer pointer = new Memory(Native.WCHAR_SIZE * (value.length() + 1));
-                pointer.setWideString(0, value);
-
-                reference.setValue(pointer);
-
-                return 0;
-            }
-        );
-
-        collection.add(new AutomationElement(elem));
-
-        when(element.findAll(any(), any())).thenReturn(collection);
-
-        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
-        UIAutomation instance = new UIAutomation(mocked_automation);
-
-        ItemContainer container = Mockito.mock(ItemContainer.class);
-
-        AutomationNUIPane pane = new AutomationNUIPane(element, container, instance);
-
-        pane.getNetUIHWND(0);
-
-        Mockito.verify(element, atLeastOnce()).findAll(any(), any());
     }
 }
