+++ src/test/java/mmarquee/automation/controls/ribbon/AutomationRibbonBarTest.java
@@ -15,28 +15,16 @@
  */
 package mmarquee.automation.controls.ribbon;
 
-import com.sun.jna.Memory;
-import com.sun.jna.Native;
-import com.sun.jna.Pointer;
-import com.sun.jna.ptr.PointerByReference;
 import mmarquee.automation.AutomationElement;
-import mmarquee.automation.ElementNotFoundException;
 import mmarquee.automation.UIAutomation;
 import mmarquee.automation.pattern.ItemContainer;
 import mmarquee.automation.uiautomation.IUIAutomation;
-import mmarquee.automation.uiautomation.IUIAutomationElement3;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import static org.junit.Assert.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.when;
 
 /**
@@ -72,63 +60,5 @@ public class AutomationRibbonBarTest {
         String name = bar.getName();
 
         assertEquals(name, "RIBBON-01");
-    }
-
-    @Test(expected = ElementNotFoundException.class)
-    @Ignore("TODO: Skip if not windows")
-    public void testGetRibbonCommandBar_Throws_Exception_When_Not_Found() throws Exception {
-        AutomationElement element = Mockito.mock(AutomationElement.class);
-
-        List<AutomationElement> collection = new ArrayList<>();
-
-        when(element.getClassName()).thenReturn(AutomationRibbonBar.CLASS_NAME);
-        when(element.findAll(any(), any())).thenReturn(collection);
-
-        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
-        UIAutomation instance = new UIAutomation(mocked_automation);
-
-        ItemContainer container = Mockito.mock(ItemContainer.class);
-
-        AutomationRibbonBar bar = new AutomationRibbonBar(element, container, instance);
-
-        bar.getRibbonCommandBar();
-
-        Mockito.verify(element, atLeastOnce()).findAll(any(), any());
-    }
-
-    @Test
-    @Ignore("TODO: Skip if not windows")
-    public void testGetRibbonCommandBar_When_Element_Is_Found() throws Exception {
-        AutomationElement element = Mockito.mock(AutomationElement.class);
-
-        when(element.getClassName()).thenReturn(AutomationRibbonBar.CLASS_NAME);
-        List<AutomationElement> collection = new ArrayList<>();
-
-        IUIAutomationElement3 elem = Mockito.mock(IUIAutomationElement3.class);
-
-        Mockito.when(elem.getCurrentClassName(any())).thenAnswer(
-                invocation -> {
-                    Object[] args = invocation.getArguments();
-                    PointerByReference reference = (PointerByReference) args[0];
-
-                    String value = "UIRibbonCommandBar";
-                    Pointer pointer = new Memory(Native.WCHAR_SIZE * (value.length() + 1));
-                    pointer.setWideString(0, value);
-
-                    reference.setValue(pointer);
-
-                    return 0;
-                }
-        );
-
-        collection.add(new AutomationElement(elem));
-
-        when(element.findAll(any(), any())).thenReturn(collection);
-
-        AutomationRibbonBar bar = new AutomationRibbonBar(element);
-
-        bar.getRibbonCommandBar();
-
-        Mockito.verify(element, atLeastOnce()).findAll(any(), any());
     }
 }
