+++ src/test/java/mmarquee/automation/controls/ribbon/AutomationRibbonBarTest2.java
@@ -32,7 +32,7 @@ import org.mockito.MockitoAnnotations;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.when;
@@ -40,17 +40,13 @@ import static org.mockito.Mockito.when;
 /**
  * Created by Mark Humphreys on 25/09/2017
  *
- * Tests for RibbonCommandBar
+ * Tests for AutomationRibbonBar.
  */
-public class AutomationRibbonCommandBarTest2 {
+public class AutomationRibbonBarTest2 {
 
     @Before
     public void setup() {
         MockitoAnnotations.initMocks(this);
-    }
-
-    static {
-        ClassLoader.getSystemClassLoader().setDefaultAssertionStatus(true);
     }
 
     @BeforeClass
@@ -62,28 +58,33 @@ public class AutomationRibbonCommandBarTest2 {
         return System.getProperty("os.name").toLowerCase().contains("windows");
     }
 
-    @Test (expected = ElementNotFoundException.class)
-    public void testGetRibbonCommandBar_Throws_Exception_When_No_WorkPane_Found() throws Exception {
+    static {
+        ClassLoader.getSystemClassLoader().setDefaultAssertionStatus(true);
+    }
+
+    @Test(expected = ElementNotFoundException.class)
+    public void testGetRibbonCommandBar_Throws_Exception_When_Not_Found() throws Exception {
         AutomationElement element = Mockito.mock(AutomationElement.class);
 
         List<AutomationElement> collection = new ArrayList<>();
 
-        when(element.getClassName()).thenReturn(AutomationRibbonCommandBar.CLASS_NAME);
+        when(element.getClassName()).thenReturn(AutomationRibbonBar.CLASS_NAME);
         when(element.findAll(any(), any())).thenReturn(collection);
 
         ItemContainer container = Mockito.mock(ItemContainer.class);
 
-        AutomationRibbonCommandBar commandBar = new AutomationRibbonCommandBar(element, container);
+        AutomationRibbonBar bar = new AutomationRibbonBar(element, container);
 
-        AutomationRibbonWorkPane workPane = commandBar.getRibbonWorkPane();
+        bar.getRibbonCommandBar();
 
         Mockito.verify(element, atLeastOnce()).findAll(any(), any());
     }
 
     @Test
-    public void testGetRibbonCommandBar_When_WorkPane_Found() throws Exception {
+    public void testGetRibbonCommandBar_When_Element_Is_Found() throws Exception {
         AutomationElement element = Mockito.mock(AutomationElement.class);
 
+        when(element.getClassName()).thenReturn(AutomationRibbonBar.CLASS_NAME);
         List<AutomationElement> collection = new ArrayList<>();
 
         IUIAutomationElement3 elem = Mockito.mock(IUIAutomationElement3.class);
@@ -93,7 +94,7 @@ public class AutomationRibbonCommandBarTest2 {
                     Object[] args = invocation.getArguments();
                     PointerByReference reference = (PointerByReference) args[0];
 
-                    String value = "UIRibbonWorkPane";
+                    String value = "UIRibbonCommandBar";
                     Pointer pointer = new Memory(Native.WCHAR_SIZE * (value.length() + 1));
                     pointer.setWideString(0, value);
 
@@ -105,12 +106,11 @@ public class AutomationRibbonCommandBarTest2 {
 
         collection.add(new AutomationElement(elem));
 
-        when(element.getClassName()).thenReturn(AutomationRibbonCommandBar.CLASS_NAME);
         when(element.findAll(any(), any())).thenReturn(collection);
 
-        AutomationRibbonCommandBar commandBar = new AutomationRibbonCommandBar(element);
+        AutomationRibbonBar bar = new AutomationRibbonBar(element);
 
-        AutomationRibbonWorkPane workPane = commandBar.getRibbonWorkPane();
+        bar.getRibbonCommandBar();
 
         Mockito.verify(element, atLeastOnce()).findAll(any(), any());
     }
